refactor(bench): unify the v1/v2 ablation records into one artifact
The two retrieval-quality campaign PRs (#37, #38) left sibling records:
ablation.{json,md} (cairn-quality-ablation/1) and ablation-v2.{json,md}
(schema /2), guarded by two test files. Per owner request they are now
ONE record: ablation.{json,md} carries the /2 document (both campaign
families, closed verdict) with the first campaign's /1 record embedded
VERBATIM under campaigns.retrieval-quality-v1 -- nothing retyped, no
rows migrated across measurement protocols (the never-cross-diff rule
survives; the legacy single-split rows stay in the history block, never
in the rows array). Original blob hashes are recorded in the artifact
and pinned by the merged guard tests (the TC-028 pin moved from the
removed sibling files onto the embedded copy and the appendix's
verbatim block); git history keeps the standalone originals at 7d9049e
and earlier. The human rendering gains a verbatim appendix of the first
campaign's md. Live pointers updated (MEASURE.md, fr003/fr004/ladder-v2
runner docstrings, CHANGELOG, specs INDEX); measured JSON artifacts and
the closed spec dirs keep their historical strings untouched. Dataset
labels (DS-v1/DS-v2) and family names are NOT campaign versioning --
they name datasets and measurement protocols and stay. Guard tests: 7
passed in one file; full suite 1913 passed, 1 skipped.
