+++ specs/INDEX.md
@@ -1,4 +1,4 @@
 # Specs index
 - [benchmark-datasource](benchmark-datasource/spec.md) — done (20/20 tasks, merged via #35 2026-08-16)
 - [retrieval-quality](retrieval-quality/spec.md) — done (24/24 tasks, merged via #37 2026-08-16; SC-1 shortfall documented in benchmarks/quality/ablation.md)
-- [retrieval-quality-v2](retrieval-quality-v2/spec.md) — done (24/24 tasks 2026-08-17; k-fold + DS-v2 evidence base; 3 candidates cleared the DS-v1 guard, all refuted zero-shot transfer; document branch — no ship, verdict in benchmarks/quality/ablation-v2.md)
+- [retrieval-quality-v2](retrieval-quality-v2/spec.md) — done (24/24 tasks 2026-08-17; k-fold + DS-v2 evidence base; 3 candidates cleared the DS-v1 guard, all refuted zero-shot transfer; document branch — no ship, verdict in benchmarks/quality/ablation.md)
